duckdb: file list via glob
