+++ vortex-duckdb-ext/Cargo.toml
@@ -22,7 +22,9 @@ crate-type = ["staticlib", "cdylib"]
 [dependencies]
 arrow-array = { workspace = true }
 bitvec = { workspace = true }
+crossbeam-queue = { workspace = true }
 duckdb = { version = "1.3.0", features = ["vtab-full"] }
+glob = { workspace = true }
 itertools = { workspace = true }
 num-traits = { workspace = true }
 tempfile = { workspace = true }
